move code into istio package
